Add basic test structure
