+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 
 [dependencies]
 actix-web = "3.0.0"
+actix-rt = "1.1.1"
 anyhow = "1.0.0"
 async-trait = "0.1.0"
 chrono = "0.4.0"
@@ -14,6 +15,7 @@ chrono-tz = "0.5.0"
 convert_case = "0.4.0"
 dotenv = "0.15.0"
 futures = "0.3.0"
+json = "0.12.4"
 regex = "1.0.0"
 serde = "1.0.0"
 serde_json = "1.0.0"
